+++ android/src/main/kotlin/com/fluttercandies/photo_manager/core/utils/IDBUtils.kt
@@ -579,7 +579,7 @@ interface IDBUtils {
 
     fun moveToGallery(context: Context, assetId: String, galleryId: String): AssetEntity
 
-    fun updateDateTaken(context: Context, assetId: String, timestamp: Int): Boolean
+    fun updateDateTaken(context: Context, assetId: String, timestamp: Long): Boolean
 
     fun getSomeInfo(context: Context, assetId: String): Pair<String, String?>?
